+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This update introduces minor bug fixes & improvements
  - Added new dynamic constant `%BUILD_OUTPUT_PATH%` which can be used as `%BUILD_OUTPUT_PATH%:release` to show the
    output path of a specific build configuration
  - Refactor CI templates to support dynamic build targets
+ - Added template `phpexe` & `phpgz` for generating executable binaries and compressed executables for the project
 
 ### Changed
  - Refactor phpmake template to support dynamic build targets